backend/requirements.txt
-i https://pypi.org/simple
alembic==1.12.1; python_version >= '3.7'
anyio==3.7.1; python_version >= '3.7'
click==8.1.7; python_version >= '3.7'
fastapi==0.104.1; python_version >= '3.8'
greenlet==3.0.1; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==21.2.0; python_version >= '3.5'
h11==0.14.0; python_version >= '3.7'
idna==3.6; python_version >= '3.5'
mako==1.3.0; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
orjson==3.9.10; python_version >= '3.8'
packaging==23.2; python_version >= '3.7'
psycopg==3.1.13; python_version >= '3.7'
psycopg-binary==3.1.13; python_version >= '3.7'
psycopg2==2.9.9; python_version >= '3.7'
psycopg2-binary==2.9.9; python_version >= '3.7'
pydantic==1.10.13; python_version >= '3.7'
python-decouple==3.8
pytz==2023.3.post1
sniffio==1.3.0; python_version >= '3.7'
sqlalchemy==2.0.23; python_version >= '3.7'
sqlmodel==0.0.12; python_version >= '3.7' and python_version < '4.0'
starlette==0.27.0; python_version >= '3.7'
typing-extensions==4.8.0; python_version >= '3.8'
uvicorn==0.24.0.post1; python_version >= '3.8'
